+++ CHANGES.rst
@@ -2,7 +2,19 @@ Changelog of threedi-modelchecker
 =================================
 
 
-1.0.2 (unreleased)
+2.0.2 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
+2.0.1 (2023-03-20)
+------------------
+
+- Pin minor version for threede-schema dependency.
+
+
+2.0.0 (2023-03-20)
 ------------------
 
 - Add warning 108: the crest_level of a weir or orifice cannot be lower than
@@ -28,6 +40,12 @@ Changelog of threedi-modelchecker
 
 - Add check 326: this gives an info message if a record exists in the simple_infiltration
   table, but is not referenced from the global settings.
+
+- Add check 66: this raises a warning if a pumpstation empties its storage area in less than one timestep.
+
+- Add check 1205 to make sure that a timeseries is not an empty string.
+
+- Add checks 1206 to confirm that the timesteps in all boundary condition timesteps are the same.
 
 
 1.0.1 (2023-02-02)